Move play again logic into the menu
