+++ src/main/java/com/samhan/ui/ConsoleOptionMenu.java
@@ -4,10 +4,13 @@ import com.samhan.BoardType;
 import com.samhan.PlayerType;
 
 import java.io.*;
+import java.util.Arrays;
 import java.util.Map;
 
 public class ConsoleOptionMenu implements OptionMenu {
     private static final String INVALID_SELECTION = "Invalid Selection.";
+    private static final String PLAY_AGAIN = "Do you want to play again? ";
+    private static final String INVALID_ENTRY = "Invalid Entry. [Y]es or [N]o";
     private final PrintStream output;
     private String promptMessage;
     private Map<String, String> options;
@@ -40,6 +43,12 @@ public class ConsoleOptionMenu implements OptionMenu {
         return PlayerType.getType(playerSelection);
     }
 
+    @Override
+    public boolean doPlayAgain() {
+        output.print(PLAY_AGAIN);;
+        return readSelectionYesNo();
+    }
+
 
     private String getPrompt(String playerNumber) {
         return String.format("Enter a Player%s selection", playerNumber);
@@ -58,6 +67,28 @@ public class ConsoleOptionMenu implements OptionMenu {
             displayError(INVALID_SELECTION);
             return getSelection(promptMessage, options);
         }
+    }
+
+    private boolean readSelectionYesNo() {
+        String selection = userInput.readInput();
+        if (validYesNo(selection)) {
+            return isYes(selection);
+        } else {
+            displayInputError();
+            return doPlayAgain();
+        }
+    }
+
+    private void displayInputError() {
+        output.println(INVALID_ENTRY);
+    }
+
+    private boolean validYesNo(String selection) {
+        return Arrays.asList("y", "n", "yes", "no").contains(selection.toLowerCase());
+    }
+
+    private boolean isYes(String selection) {
+        return Arrays.asList("y", "yes").contains(selection.toLowerCase());
     }
 
     private void displayError(String errorMessage) {
