--- src/main/java/com/samhan/ui/ConsolePlayAgainInput.java
@@ -1,51 +1,0 @@
-package com.samhan.ui;
-
-import java.io.*;
-import java.util.Arrays;
-
-public class ConsolePlayAgainInput {
-    private static final String ANSI_CLS = "\u001b[2J";
-    private static final String ANSI_HOME = "\u001b[H";
-    private static final String WELCOME_MESSAGE = "Welcome to TicTcToe\n\nTime to set up!\n";
-    private static final String FAREWELL_MESSAGE = "\nThanks for playing!\n\nBye!";
-    private static final String PLAY_AGAIN = "Do you want to play again? ";
-    private static final String INVALID_ENTRY = "Invalid Entry. [Y]es or [N]o";
-    private final PrintStream output;
-    private final UserInput userInput;
-
-    public ConsolePlayAgainInput(UserInput userInput, PrintStream output) {
-        this.userInput = userInput;
-        this.output = output;
-    }
-
-    public boolean doPlayAgain() {
-        displayPrompt();
-        return readSelection();
-    }
-
-    private void displayPrompt() {
-        output.print(PLAY_AGAIN);
-    }
-
-    private boolean readSelection() {
-        String selection = userInput.readInput();
-        if (valid(selection)) {
-            return isYes(selection);
-        } else {
-            displayInputError();
-            return doPlayAgain();
-        }
-    }
-
-    private void displayInputError() {
-        output.println(INVALID_ENTRY);
-    }
-
-    private boolean valid(String selection) {
-        return Arrays.asList("y", "n", "yes", "no").contains(selection.toLowerCase());
-    }
-
-    private boolean isYes(String selection) {
-        return Arrays.asList("y", "yes").contains(selection.toLowerCase());
-    }
-}
